This repository consists of Verilog implementation of Authenticated Encryption Decryption
This design consists of paper implementation of TIAOXIN-346, a CAESAR
entry developed by Ivica Nicolic. This project is carried for the
Hardware Implementation which is not listed in the paper submitted in
CAESAR. Algorithm consists of AES as a part of encryption and decryption
and authenticity is added through Tag (sequence). Algorithm is available
on CAESAR website.